refactor: simplify file orgqanization
