--- internal/constant.go
@@ -1,16 +1,0 @@
-package internal
-
-// App information constants
-const APP_NAME = "Go Phrase"
-const APP_USAGE = "CLI for generating secure, memorable passwords"
-const APP_VERSION = "v1.0.0 | release: 5.10.20"
-const APP_USAGETEXT = "gophrase gen [word count] [word list] [--flags]"
-
-// Static asset constants
-const EFF_LARGE = "eff_large_wordlist.json"
-const EFF_SHORT_1 = "eff_short_wordlist_1.json"
-const EFF_SHORT_2 = "eff_short_wordlist_2_0.json"
-const REINHOLD = "reinhold_wordlist.json"
-const CHARACTERS = "special_characters.json"
-const LIST_OPTIONS = "list_options.txt"
-const DEFAULTS = "defaults.json"
