+++ src/assets/special_characters.json
