+++ src/config/config.go
@@ -4,10 +4,18 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/gobuffalo/packr/v2"
-	"github.com/gophrase/internal"
 	"io/ioutil"
 	"log"
 )
+
+// Static asset constants
+const EFF_LARGE = "eff_large_wordlist.json"
+const EFF_SHORT_1 = "eff_short_wordlist_1.json"
+const EFF_SHORT_2 = "eff_short_wordlist_2_0.json"
+const REINHOLD = "reinhold_wordlist.json"
+const CHARACTERS = "special_characters.json"
+const LIST_OPTIONS = "list_options.txt"
+const DEFAULTS = "defaults.json"
 
 type Default struct {
 	WordCount int    `json:"wordCount"`
@@ -18,7 +26,7 @@ var Assets = packr.New("assets", "../../assets")
 var Defaults = new(Default)
 
 func SetConfigDefaults(config *Default, count int, list string) {
-	fileLocation := internal.DEFAULTS //TODO create better implementation
+	fileLocation := DEFAULTS // TODO create better implementation
 	config.WordCount = count
 	config.WordList = list
 	file, err := json.MarshalIndent(config, "", "\t")
@@ -32,7 +40,7 @@ func SetConfigDefaults(config *Default, count int, list string) {
 }
 
 func LoadConfigDefaults() {
-	fileLocation, err := Assets.Find(internal.DEFAULTS)
+	fileLocation, err := Assets.Find(DEFAULTS)
 	if err != nil {
 		log.Fatal(err)
 	}
