+++ src/embedded_licenses/github.com/markbates/safe/LICENSE
