+++ src/embedded_licenses/github.com/shurcooL/sanitized_anchor_name/LICENSE
