+++ src/embedded_licenses/github.com/sirupsen/logrus/LICENSE
