+++ src/entropy/entropy_test.go
