+++ src/go.mod
@@ -1,10 +1,9 @@
-module github.com/gophrase
+module github.com/jmillerv/gophrase
 
-go 1.13
+go 1.18
 
 require (
 	github.com/gobuffalo/packr/v2 v2.8.0
 	github.com/nbutton23/zxcvbn-go v0.0.0-20180912185939-ae427f1e4c1d
 	github.com/urfave/cli/v2 v2.1.1
-	golang.org/x/tools v0.0.0-20200308013534-11ec41452d41
 )
